configure generate file to give more emotions as prompts to OpenAI (#7)
configure generate file to give more emotions as prompts to OpenAI (#7)

What changed?
==================
This commit tweaks a few of the completion APIs settings and redefines some of the moods in the generatePrompt function to give more moods so the OpenAI can process user requests better and with greater accuracy.

Why?
==================
This commit gives users better song suggestions according to the user submitted mood. It also has a greater amount of actual songs instead of AI generated fake song titles with real artists.