add connection to youtube search api (#8)
add connection to youtube search api (#8)

What changed?
===============
This commit adds a connection to the Youtube API. A GET request is sent to the API which sends back an object containing data on a Youtube video.

Why?
===============
This commit allows the AI to send a song as a response via a Youtube video which users can then click on to direct them to the video. The video link is an external link as Youtube does not allow music to be embedded due to copyright concerns.