fix order of results to store from newest to oldest (#10)
fix order of results to store from newest to oldest (#10)

What changed?
================
This commit fixes the order of the results state to have the newest result be at the front of the array, index 0, and the oldest to be at the end of the array. Also, rendering of the results array was reversed.

Why?
===============
This order of the results state was changed to allow easy access to the latest message result from a prompt being sent by a user. Rendering was reversed as I want the UI to be like a chat message having the newest message at the bottom and the view to scroll automatically as subsequent messages are sent through.