add user preset prompts and refactor (#21)
add user preset prompts and refactor (#21)

What changed?
===============
This commit adds the user preset prompts feature to the chat bar. This commit also has some refactoring.

Why?
==============
This commit allows users to choose from a pre-determined list of prompts for the user to select from. This allows users who might not know what to type an idea, and doubles as a tutorial for users.