fix types, removing any types (#22)
fix types, removing any types (#22)

What changed?
===============
This commit adds fixes to the typing of props and functions. It ensures that variables or function arguments do not use the any type as much as possible.

Why?
===============
This commit ensures that contributors to the code base has an easier time with typings.